requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
python-dotenv>=1.0.0
openai>=1.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
PyMySQL>=1.1.0
requests>=2.31.0
openpyxl>=3.1.0
Pillow>=10.2.0
